--- apps/client/src/pages/home/constants/menu.ts
@@ -1,5 +1,0 @@
-export enum TAB_MENU {
-  HOME = 0,
-  TIMETABLE = 1,
-  MY_HISTORY = 2,
-}
