+++ apps/client/src/pages/home/constants/tab.ts
@@ -1,0 +1,11 @@
+export enum TAB_MENU {
+  HOME = 0,
+  TIMETABLE = 1,
+  MY_HISTORY = 2,
+}
+
+export const enum HOME_CATEGORY_TAB {
+  TICKETING = '티켓 오픈',
+  SUGGEST_PERFORMANCE = '추천 공연',
+  PLAYLIST = '플레이리스트',
+}
