+++ apps/client/src/shared/components/navigation-tabs.tsx
@@ -4,7 +4,7 @@ import { Navigation } from '@confeti/design-system';
 
 import { routePath } from '@shared/router/path';
 
-import { TAB_MENU } from '@pages/home/constants/menu';
+import { TAB_MENU } from '@pages/home/constants/tab';
 
 interface Props {
   defaultActiveTab: TAB_MENU;
